requirements.txt
streamlit>=1.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
openai>=1.0.0
crawl4ai>=0.8.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
playwright-stealth>=1.0.0
thefuzz>=0.22.0
python-Levenshtein>=0.25.0
requests>=2.31.0
tabulate>=0.9.0